IPEX support for Torch 2.1 and fix dtype erros
